server/requirements.txt
fastapi
uvicorn
sqlalchemy
python-multipart
pyjwt
passlib[bcrypt]
pytesseract
Pillow
langchain
langchain-huggingface
langchain-community
